add jasperreport

+++ ciremai/middleware/views.py
@@ -13,6 +13,9 @@ from django.http import HttpResponse
 from . import models,tables,filters,forms
 from billing.models import Orders,OrderTests,Tests
 
+import os
+from pyreportjasper import JasperPy
+
 
 class FilteredSingleTableView(SingleTableView):
     filter_class = None
@@ -30,11 +33,66 @@ class FilteredSingleTableView(SingleTableView):
     
 
 def show_all_orders(request):
-    orders = Orders.objects.all()
-    ordertable = tables.OrderResultTable(orders)
+    data = Orders.objects.all()
+    if request.GET.get('order_date'):
+        d_range = request.GET.get('order_date')
+        #04/01/2018 - 04/30/2018
+        start_date = d_range[:10]
+        end_date = d_range[13:23]
+        
+        data = data.filter(order_date__range=[start_date,end_date] )
+    
+    
+    if request.GET.get('number'):
+        #print request.GET.get('number')
+        data = data.filter(number__contains=request.GET.get('number'))
+    
+    if request.GET.get('patient__patient_id'):
+        #print request.GET.get('patient__patient_id')
+        data = data.filter(patient__patient_id__contains=request.GET.get('patient__patient_id'))
+        
+    if request.GET.get('patient__name'):
+        #print request.GET.get('patient__name')
+        data = data.filter(patient__name__contains=request.GET.get('patient__name'))
+
+    filter = filters.OrderFilter(request.GET,queryset=data)
+
+    
+    ordertable = tables.OrderResultTable(data)
     ordertable.paginate(page=request.GET.get('page', 1), per_page=10)
+    
+    RequestConfig(request).configure(ordertable)
+    
     tempate = 'middleware/list_orders.html'
-    context = {'ordertable':ordertable}
+    context = {'ordertable':ordertable,'filter':filter}
+    return render(request,tempate,context)
+
+def order_results_history(request,pk):
+    order = Orders.objects.get(pk=pk)
+    data = models.HistoryOrders.objects.filter(order_id=pk)
+    
+    if request.GET.get('test'):
+        print request.GET.get('test')
+        data = data.filter(test=request.GET.get('test'))
+        
+    if request.GET.get('action_user'):
+        print request.GET.get('action_user')
+        data = data.filter(action_user__contains=request.GET.get('action_user'))
+        
+    if request.GET.get('action_code'):
+        print request.GET.get('action_code')
+        data = data.filter(action_code__contains=request.GET.get('action_code'))
+        
+    if request.GET.get('action_text'):
+        print request.GET.get('action_text')
+        data = data.filter(action_text__contains=request.GET.get('action_text'))
+    
+    filter = filters.OrderHistoryFilter(request.GET,queryset=data)
+    
+    orderhist = tables.OrderHistoryTable(data)
+    orderhist.paginate(page=request.GET.get('page', 1), per_page=20)
+    tempate = 'middleware/order_history.html'
+    context = {'order':order,'orderhist':orderhist,'filter':filter}
     return render(request,tempate,context)
 
 def order_results_validate(request,pk):
@@ -44,11 +102,29 @@ def order_results_validate(request,pk):
 
 def order_results_techval(request,pk):
     if request.user.is_authenticated():
+        # create history
+        tech_val = models.OrderResults.objects.filter(order_id=pk,validation_status=1).values('test_id')
+        for tes in tech_val:
+            test_id = tes['test_id']
+            test = Tests.objects.get(pk=test_id)
+            act_txt = 'Test %s technical validated' % (test)
+            his_order = models.HistoryOrders(order_id=pk,test=test,action_code='TECHVAL',action_user=str(request.user),action_date=datetime.now(),action_text=act_txt)
+            his_order.save()
+        
+        # update
         order_res = models.OrderResults.objects.filter(order_id=pk,validation_status=1).update(validation_status=2,techval_user=str(request.user),techval_date=datetime.now())
     return redirect('order_results', pk=pk)
 
 def order_results_medval(request,pk):
     if request.user.is_authenticated():
+        # create history
+        med_val = models.OrderResults.objects.filter(order_id=pk,validation_status=2).values('test_id')
+        for tes in med_val:
+            test_id = tes['test_id']
+            test = Tests.objects.get(pk=test_id)
+            act_txt = 'Test %s medical validated' % (test)
+            his_order = models.HistoryOrders(order_id=pk,test=test,action_code='MEDVAL',action_user=str(request.user),action_date=datetime.now(),action_text=act_txt)
+            his_order.save()
         order_res = models.OrderResults.objects.filter(order_id=pk,validation_status=2).update(validation_status=3,medval_user=str(request.user),medval_date=datetime.now())
     return redirect('order_results', pk=pk)
 
@@ -70,11 +146,57 @@ def order_results_repeat(request,pk):
         order_result.medval_date = None
         order_result.patologi_mark = None
         order_result.save()
-        print tes
         order_res = models.OrderResults.objects.filter(order_id=pk,validation_status=1).update(validation_status=3,medval_user=str(request.user),medval_date=datetime.now())
+        
+         # create history
+        act_txt = 'Result %s repeated' % (tes)
+        his_order = models.HistoryOrders(order=order,test=tes,action_code='REPEAT',action_user=str(request.user),action_date=datetime.now(),action_text=act_txt)
+        his_order.save()
     return redirect('order_results', pk=pk)
 
 def order_results_print(request,pk):
+    order = Orders.objects.get(pk=pk)
+    input_file = settings.RESULT_REPORT_FILE
+    ts = datetime.today().strftime('%Y%m%d%H%M%S')
+    parameters ={'ORDER_ID': pk}
+    output = settings.MEDIA_ROOT+'\\report\\'+str(order.number)+'_'+ts
+    con = settings.JASPER_CONN
+    jasper = JasperPy()
+    jasper.process(
+        input_file,
+        output_file=output,
+        format_list=["pdf"],
+        parameters=parameters,
+        db_connection=con,
+        locale='en_US'
+    )
+    
+    base_url =  request.build_absolute_uri('/')[:-1].strip("/")
+    url_pdf = base_url+'/media/report/'+str(order.number)+'_'+ts+'.pdf'
+    
+    # save report URL
+    oe, _created = models.OrderExtended.objects.get_or_create(order_id=pk)
+
+    oe.result_pdf_url = url_pdf
+    oe.save()
+    
+    # set validation printed
+    if request.user.is_authenticated():
+        order_res = models.OrderResults.objects.filter(order_id=pk,validation_status=3).update(validation_status=4,print_user=str(request.user),print_date=datetime.now())
+    
+    
+
+    template = 'middleware/result_pdf_preview.html'
+    context = {'order':order,'url_pdf' : url_pdf}
+    return render(request,template,context)
+    
+def order_result_report(request,pk):
+    
+    template = 'middleware/result_pdf_preview.html'
+    context = {'order':order,'url_pdf' : url_pdf}
+    return render(request,template,context)
+
+def order_results_print2(request,pk):
     order = Orders.objects.get(pk=pk)
     response = HttpResponse(content_type='application/pdf')
     response['Content-Disposition'] = 'filename="somefilename.pdf"'
@@ -148,6 +270,7 @@ def order_results_print2(request,pk):
     return render(request,template,context)
 
 def order_results(request,pk):
+    order = Orders.objects.get(pk=pk)
     if request.method == 'POST': 
         for p_tes in request.POST:
             if p_tes.startswith('test_'):
@@ -163,23 +286,29 @@ def order_results(request,pk):
                     except:
                         pass
                     
-                    if not cr_res_a == request.POST.get( p_tes, ''):           
-                        o_result = models.Results(order=o_order,test=o_test,alfa_result=request.POST.get( p_tes, ''))
-                        o_result.save()
-                        # try delete existing result
-                        try:
-                            o_orderresult = models.OrderResults.objects.get(order=o_order,test=o_test)
-                            o_orderresult.delete()
-                        except models.OrderResults.DoesNotExist,e:
-                            pass
-                        # try get unit
-                        s_unit = ''
-                        try:
-                            test_unit = models.TestParameters.objects.get(test=o_test)
-                            s_unit = test_unit.unit
-                        except models.TestParameters.DoesNotExist,e:
-                            pass
-                        o_orderresult = models.OrderResults.objects.get_or_create(order=o_order,test=o_test,result=o_result,validation_status=1,unit=s_unit)
+                    if not cr_res_a == request.POST.get( p_tes, ''):
+                        if request.user.is_authenticated():      
+                            o_result = models.Results(order=o_order,test=o_test,alfa_result=request.POST.get( p_tes, ''))
+                            o_result.save()
+                            # try delete existing result
+                            try:
+                                o_orderresult = models.OrderResults.objects.get(order=o_order,test=o_test)
+                                o_orderresult.delete()
+                            except models.OrderResults.DoesNotExist,e:
+                                pass
+                            # try get unit
+                            s_unit = ''
+                            try:
+                                test_unit = models.TestParameters.objects.get(test=o_test)
+                                s_unit = test_unit.unit
+                            except models.TestParameters.DoesNotExist,e:
+                                pass
+                            o_orderresult = models.OrderResults.objects.get_or_create(order=o_order,test=o_test,result=o_result,validation_status=1,unit=s_unit)
+                            
+                            # create history
+                            act_txt = 'Result %s set for analyt %s ' % (request.POST.get( p_tes, ''),o_test)
+                            his_order = models.HistoryOrders(order=o_order,test=o_test,action_code='RESENTRY',action_user=str(request.user),action_date=datetime.now(),action_text=act_txt)
+                            his_order.save()
     else:
         order_tests = OrderTests.objects.filter(order_id=pk)
         #print order_tests
@@ -258,9 +387,10 @@ def order_results(request,pk):
                                                                            'result__instrument__name',
                                                                            'techval_user',
                                                                            'medval_user').order_by('test__test_group__sort','test__sort')
-    print ordertests.query
+    # save report URL
+    oe, _created = models.OrderExtended.objects.get_or_create(order_id=pk)
     tempate = 'middleware/order_results.html'
-    context = {'orders':orders,'ordertests':ordertests}
+    context = {'order':order,'orders':orders,'ordertests':ordertests}
     return render(request,tempate,context)
